Parse to float to no truncate position fields

+++ webapp/public/javascripts/angular/data-series/registration.js
@@ -575,7 +575,7 @@ angular.module('terrama2.dataseries.registration', [
             if (isNaN(fmt[k]) || typeof fmt[k] == "boolean") {
               output[k] = fmt[k];
             } else {
-              output[k] = parseInt(fmt[k]);
+              output[k] = parseFloat(fmt[k]);
             }
           }
         }
